--- src/app/ImageViewerScreen.tsx
@@ -1,46 +1,0 @@
-import {StyleSheet, View} from 'react-native'
-import {
-  useSafeAreaInsets,
-  type EdgeInsets,
-} from 'react-native-safe-area-context'
-import type {NavigationProps} from '@/app/navigation/types'
-import type {Theme} from '@/themes/themes'
-import {HeaderBackButton} from '@/components/features/header/HeaderBackButton'
-import {ImageViewer} from '@/components/features/image-viewer/ImageViewer'
-import {Screen} from '@/components/features/screen/Screen'
-import {useNavigation} from '@/hooks/navigation/useNavigation'
-import {HomeModalName} from '@/modules/home/routes'
-import {useThemable} from '@/themes/useThemable'
-
-type Props = NavigationProps<HomeModalName.imageViewer>
-
-export const ImageViewerScreen = ({route}: Props) => {
-  const {goBack} = useNavigation()
-  const {top} = useSafeAreaInsets()
-
-  const styles = useThemable(createStyles(top))
-
-  return (
-    <Screen
-      scroll={false}
-      testID="ImageViewerScreen"
-      withBottomInset={false}
-      withTopInset={false}>
-      <View style={styles.backButtonContainer}>
-        <HeaderBackButton onPress={goBack} />
-      </View>
-
-      <ImageViewer {...route.params} />
-    </Screen>
-  )
-}
-
-const createStyles = (top: EdgeInsets['top']) => (theme: Theme) =>
-  StyleSheet.create({
-    backButtonContainer: {
-      position: 'absolute',
-      top: theme.size.spacing.md + top,
-      left: theme.size.spacing.md,
-      zIndex: theme.z.overlay,
-    },
-  })
